resolving conflict
