+++ http/index.ts
@@ -63,22 +63,3 @@ export const makePayment = async (selectedPaymentMethod: string) => {
     throw new Error('Please select a payment method before making the payment.');
   }
 };
-
-export const guestSignup = async (props: { email: string; firstName: string; lastName: string; password: string }) => {
-  const $http = axios.create({
-    baseURL: AUTH_HTTP_URL,
-    timeout: 30000,
-    headers: {
-      'Content-Type': 'application/json; charset=UTF-8',
-    },
-  });
-
-  try {
-    const res = await $http.post('/api/auth/signup', props);
-    console.log(res?.data);
-    return res?.data;
-  } catch (e: any) {
-    console.log(e);
-    return e.response.data ?? { message: e.message };
-  }
-};
